+++ src/models/applications.rs
@@ -10,7 +10,7 @@ use crate::schema::{application_reviews, applications};
 #[diesel(check_for_backend(diesel::mysql::Mysql))]
 pub struct Application {
     pub id: u32,
-    pub user_id: u32,
+    pub user_id: u64,
     pub accepted: bool,
     pub background: String,
     pub motivation: String,
@@ -22,7 +22,7 @@ pub struct Application {
 #[derive(Debug, Insertable, AsChangeset)]
 #[diesel(table_name = applications)]
 pub struct ApplicationModel<'a> {
-    pub user_id: u32,
+    pub user_id: u64,
     pub accepted: bool,
     pub background: &'a str,
     pub motivation: &'a str,
@@ -35,14 +35,14 @@ pub struct ApplicationModel<'a> {
 #[diesel(check_for_backend(diesel::mysql::Mysql))]
 pub struct ApplicationReview {
     pub id: u32,
-    pub reviewer_id: u32,
+    pub reviewer_id: u64,
     pub application_id: u32,
 }
 
 #[derive(Debug, Insertable, AsChangeset)]
 #[diesel(table_name = application_reviews)]
 pub struct ApplicationReviewModel {
-    pub reviewer_id: u32,
+    pub reviewer_id: u64,
     pub application_id: u32,
 }
